[change] tank1 code to work with server and lcd display

+++ arduino/src/sensors/sensor.cpp
@@ -28,4 +28,4 @@ void Sensor::setSensorType(sensor_type _type) { type = _type; }
 int Sensor::getInputPin() { return inputPin; }
 void Sensor::setInputPin(int _inputPin) { inputPin = _inputPin; }
 
-int Sensor::readInput() { return 0; }
+long Sensor::readInput() { return 0; }
